test: split cookie concurrency coverage
